[API] feat: Add errorHandling and app file (#27)
* feat: Add errorHandling and app file

* chore(server): auto-format TypeScript code via CI
